// src/payload-types.ts
export interface Product {
  id: number;
  pid?: string | null;
  title: string;
  currency?: string | null;
  visible?: boolean | null;
  /**
   * Choose where this product should be available to customers.
   */
  salesChannels?: ('all' | 'onlineStore' | 'pos' | 'mobileApp')[] | null;
  source?: ('manual' | 'cj') | null;
  description?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  collections?: (number | Collection)[] | null;
  handle?: string | null;
  /**
   * Choose the options for this product.
   */
  variantOptions?:
    | {
        option: string;
        /**
         * (press enter to add multiple values)
         */
        value: string[];
        id?: string | null;
      }[]
    | null;
  variants: {
    vid?: string | null;
    imageUrl?: string | null;
    gallery?: (number | Media)[] | null;
    price: number;
    originalPrice?: number | null;
    options?:
      | {
          option: string;
          value: string;
          id?: string | null;
        }[]
      | null;
    id?: string | null;
  }[];
  /**
   * Add additional product info such as care instructions, materials, or sizing notes.
   */
  customFields?:
    | {
        name: string;
        value?: string | null;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "collections".
 */
export interface Collection {
  id: number;
  title: string;
  imageUrl?: string | null;
  handle?: string | null;
  description?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  products?: {
    docs?: (number | Product)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}